+++ web/src/server_events_dispatch.js
@@ -200,6 +200,7 @@ export function dispatch_normal_event(event) {
                 allow_edit_history: noop,
                 allow_message_editing: noop,
                 edit_topic_policy: noop,
+                can_resolve_topics_group: noop,
                 user_group_edit_policy: noop,
                 avatar_changes_disabled: settings_account.update_avatar_change_display,
                 bot_creation_policy: settings_bots.update_bot_permissions_ui,
@@ -296,7 +297,10 @@ export function dispatch_normal_event(event) {
                                     stream_settings_ui.update_stream_privacy_choices(key);
                                 }
 
-                                if (key === "edit_topic_policy") {
+                                if (
+                                    key === "edit_topic_policy" ||
+                                    key === "can_resolve_topics_group"
+                                ) {
                                     message_live_update.rerender_messages_view();
                                 }
                             }